requirements.txt
gTTS==2.5.1
Pillow==10.2.0
playsound==1.3.0
rembg==2.0.55
streamlit==1.31.1
